+++ src/readFileTree.ts
@@ -1,119 +1,76 @@
-import { Ignore } from "./ignore";
-import { readdirSync, Dirent, statSync } from "fs";
-import { joinPaths, winPathToUnixPath } from "./path";
+import { Dirent } from "fs";
+import { Dir, FSObject } from "./filesystem";
+import { IgnoreCache } from "./ignore";
+import upath from "upath";
 
-type ReadFunc = (filename: string, depth: number) => string;
-
-function fileToStr(filename: string, depth: number) {
-  return `${"\t".repeat(depth)}${filename}`;
+enum Connector {
+  TAB = "    ",
+  LINE = "│   ",
+  T = "├── ",
+  L = "└── ",
+  T_IGNORED = "├─] ",
+  L_IGNORED = "└─] ",
 }
 
-function dirToStr(filename: string, depth: number) {
-  return `${"\t".repeat(depth)}${filename}/`;
-}
+class Out {
+  buffer: string;
 
-enum branch {
-  none = " ",
-  between = "│",
-  middleConnected = "├",
-  lastConnected = "└",
-}
-
-export class Tree {
-  buffer: string = "";
-  constructor(public ignore: Ignore, public selectedPath: string) { }
-  readFile(
-    filename: string,
-    branches: branch[],
-    isDir: boolean,
-    ignored: boolean = false
-  ) {
-    let branchesString = branches.reduce((a, b) => {
-      const connection =
-        a in [branch.middleConnected, branch.lastConnected] ? "──" : "  ";
-      return a + connection + b;
-    }, "");
-    let branchesStringSuffix = "─ ";
-    // if (!isDir) {
-    //   branchesStringSuffix = "──>";
-    // }
-    if (ignored) {
-      // branchesString = branchesString.replace(branch.lastConnected, "┗");
-      // branchesString = branchesString.replace(branch.middleConnected, "┣");
-      branchesStringSuffix = "─]";
-    }
-    if (branches.length === 0) {
-      branchesStringSuffix = "";
-    }
-    this.buffer += `${branchesString}${branchesStringSuffix}${filename}${isDir ? "/" : ""
-      }${ignored ? " (ignored)" : ""}\n`;
+  constructor() {
+    this.buffer = "";
   }
-  read(): string {
-    this.recTree(this.selectedPath, this.selectedPath, [], true, true);
-    return this.buffer;
+
+  println(str: string) {
+    this.buffer += str + "\n";
   }
-  recTree(
-    path: string,
-    name: string,
-    branches: branch[],
-    isDir: boolean,
-    isLast: boolean
-  ) {
-    if (name === ".git") {
-      if (branches.length === 0) {
-        this.readFile(name, branches, isDir);
-      } else {
-        return;
-      }
+}
+
+const defaultSkipAliases: string[] = [".git"];
+class DirTreeReader {
+  #out: Out;
+  #ignore: IgnoreCache;
+  constructor(path: string, ignore: IgnoreCache) {
+    this.#out = new Out();
+    this.#ignore = ignore;
+    this.#recReadDir(Dir.fromPath(path), "", true);
+  }
+
+  #recReadDir(obj: FSObject, context: string, isLast: boolean): void {
+    if (defaultSkipAliases.includes(obj.alias)) {
+      return;
     }
-    if (this.ignore.ignores(path)) {
-      this.readFile(name, branches, isDir, true);
+
+    let myConnector: string;
+    if (isLast) {
+      myConnector = obj.isIgnored ? Connector.L_IGNORED : Connector.L;
     } else {
-      this.readFile(name, branches, isDir);
-      if (isDir) {
-        let entries = readdirSync(path, { withFileTypes: true });
-        this.ignore.applyGitignore(path);
-        entries.sort((a, b) => {
-          // Directories first
-          if (a.isDirectory() && b.isFile()) {
-            return -1;
-          }
-          if (b.isDirectory() && a.isFile()) {
-            return 1;
-          }
+      myConnector = obj.isIgnored ? Connector.T_IGNORED : Connector.T;
+    }
 
-          // Lexicographical order
-          return a.name.localeCompare(b.name);
-        });
-        entries.map((entry, index) => {
-          const entryBranches = [...branches];
+    const normalizedAlias = upath.normalize(obj.alias);
 
-          // If dir is last entry of its parent, change branch at that level to none,
-          // otherwise change it to between
-          if (entryBranches.length !== 0) {
-            if (isLast) {
-              entryBranches[entryBranches.length - 1] = branch.none;
-            } else {
-              entryBranches[entryBranches.length - 1] = branch.between;
-            }
-          }
+    const suffix =
+      (obj instanceof Dir ? "/" : "") + (obj.isIgnored ? " (ignored)" : "");
 
-          const isChildLast = index === entries.length - 1;
-          // If child is last, add lastConnected to its branches, otherwise add middleConnected
-          if (isChildLast) {
-            entryBranches.push(branch.lastConnected);
-          } else {
-            entryBranches.push(branch.middleConnected);
-          }
-          this.recTree(
-            joinPaths(path, entry.name),
-            entry.name,
-            entryBranches,
-            entry.isDirectory(),
-            isChildLast
-          );
-        });
-      } // else is file, maybe symbolic link?
+    this.#out.println(context + myConnector + normalizedAlias + suffix);
+
+    // If obj is a directory, recursively read its children
+    if (obj instanceof Dir && !obj.isIgnored) {
+      const childConnector = isLast ? Connector.TAB : Connector.LINE;
+      const childContext = context + childConnector;
+      const children = obj.getChildren(this.#ignore);
+      children.forEach((child, index) => {
+        const isChildLast = index === children.length - 1;
+        this.#recReadDir(child, childContext, isChildLast);
+      });
     }
   }
+
+  result(): string {
+    return this.#out.buffer;
+  }
+}
+
+export function readDirTree(dirPath: string, ig: IgnoreCache): string {
+  const tree = new DirTreeReader(dirPath, ig);
+  return tree.result();
 }